+++ Shared/CMdmi.Entities/Domain/Customer/CDMA_IDENTIFICATION_TYPE.cs
@@ -20,7 +20,7 @@
 
         
         [Key]
-        public decimal CODE { get; set; }
+        public int CODE { get; set; }
         public string ID_TYPE { get; set; }
         public ICollection<CDMA_INDIVIDUAL_NEXT_OF_KIN> CdmaNextOfKins { get; private set; }
     }
